Move getNameOfEntity function to shared/.
This also adds a sortByName function.
